Make site sampling more robust
Before, we assumed that a site WOULD be chosen given a bulk/surface. But
if a surface is fully sampled already, then we want to move on to
another surface. And the same thing goes for bulks. This commit attempts
to loop through bulks/surfaces to enable this fall-back logic.